refactor: replace modular MCP files with unified loader module and update references
