--- ollama_agent/mcp/lifecycle.py
@@ -1,292 +1,0 @@
-"""MCP server initialization and cleanup routines.
-
-Loads MCP servers from a JSON config and exposes each server as a Deep Agents
-subagent descriptor.
-"""
-
-from __future__ import annotations
-
-import json
-import logging
-from copy import deepcopy
-from pathlib import Path
-from typing import Any
-
-from langchain_core.tools import BaseTool, StructuredTool
-from langchain_mcp_adapters.tools import load_mcp_tools
-
-from ..core import (
-    ModelCapabilityError,
-    ReasoningEffortValue,
-    create_ollama_chat_model,
-    ensure_model_supports_tools,
-    validate_reasoning_effort,
-)
-from .types import DEFAULT_AGENT_INSTRUCTIONS, DEFAULT_MCP_CONFIG_PATH, RunningMCPServer
-
-logger = logging.getLogger(__name__)
-
-
-async def _noop_shutdown() -> None:
-    return
-
-
-def _relax_const_fields(tools: list[BaseTool]) -> list[BaseTool]:
-    """Auto-fill ``const``-constrained parameters in MCP tool schemas.
-
-    Some MCP servers declare JSON Schema parameters with ``"const": "value"``
-    allowing only one value.  LLMs may guess a different value, causing
-    Pydantic validation errors.  This removes the ``const`` constraint from the
-    schema (so the LLM can omit the field) and forces the correct value at call
-    time.  Fully generic — no server-specific knowledge is needed.
-    """
-    result: list[BaseTool] = []
-    for tool in tools:
-        schema = getattr(tool, "args_schema", None)
-        if not isinstance(schema, dict):
-            result.append(tool)
-            continue
-
-        props = schema.get("properties", {})
-        if not isinstance(props, dict):
-            result.append(tool)
-            continue
-
-        consts: dict[str, Any] = {}
-        for fname, fschema in props.items():
-            if isinstance(fschema, dict) and "const" in fschema:
-                consts[fname] = fschema["const"]
-
-        if not consts:
-            result.append(tool)
-            continue
-
-        patched = deepcopy(schema)
-        for fname, const_val in consts.items():
-            field = patched["properties"][fname]
-            field.pop("const", None)
-            field.setdefault("default", const_val)
-
-        orig_coro = getattr(tool, "coroutine", None)
-        if callable(orig_coro):
-
-            def _make_wrapped(coro: Any, const_args: dict[str, Any]) -> Any:
-                async def _wrapped(*args: Any, **kwargs: Any) -> Any:
-                    kwargs.update(const_args)
-                    return await coro(*args, **kwargs)
-
-                return _wrapped
-
-            _wrapped_coro = _make_wrapped(orig_coro, consts)
-
-            result.append(
-                StructuredTool(
-                    name=tool.name,
-                    description=tool.description,
-                    args_schema=patched,
-                    coroutine=_wrapped_coro,
-                    response_format=getattr(tool, "response_format", "content"),
-                    metadata=getattr(tool, "metadata", None),
-                )
-            )
-        else:
-            result.append(tool)
-
-    return result
-
-
-def _get(cfg: dict[str, Any], *keys: str) -> Any:
-    return next((cfg[k] for k in keys if k in cfg), None)
-
-
-def _infer_transport(cfg: dict[str, Any]) -> str:
-    t = _get(cfg, "type", "transport") or ""
-    if isinstance(t, str) and t:
-        return t.lower()
-    if cfg.get("command"):
-        return "stdio"
-    if _get(cfg, "httpUrl", "url"):
-        return "http"
-    return ""
-
-
-def _build_connection(cfg: dict[str, Any]) -> dict[str, Any] | None:
-    transport = _infer_transport(cfg)
-    if transport in ("process", "stdio"):
-        command = cfg.get("command")
-        if not command:
-            return None
-        out: dict[str, Any] = {"transport": "stdio", "command": command}
-        if "args" in cfg:
-            out["args"] = cfg["args"]
-        for k in ("env", "cwd", "encoding", "encoding_error_handler", "session_kwargs"):
-            if k in cfg:
-                out[k] = cfg[k]
-        return out
-
-    if transport in ("http", "streamable_http", "streamable", "sse", "http_sse"):
-        url = _get(cfg, "url", "httpUrl")
-        if not url:
-            return None
-        out = {"transport": "http", "url": url}
-        for k in (
-            "headers",
-            "timeout",
-            "sse_read_timeout",
-            "terminate_on_close",
-            "session_kwargs",
-            "auth",
-        ):
-            if k in cfg:
-                out[k] = cfg[k]
-        return out
-
-    return None
-
-
-async def _init_server(
-    name: str, config: Any, default_model: str | None
-) -> RunningMCPServer | None:
-    if not isinstance(config, dict):
-        logger.warning("Skipping MCP server '%s': invalid config type", name)
-        return None
-
-    if not (connection := _build_connection(config)):
-        logger.warning("Skipping MCP server '%s': could not determine transport", name)
-        return None
-
-    raw_agent_cfg = config.get("agent", {})
-    agent_cfg = raw_agent_cfg if isinstance(raw_agent_cfg, dict) else {}
-    model = agent_cfg.get("model") or default_model
-    if not model:
-        logger.error("Skipping MCP server '%s': missing model", name)
-        return None
-
-    try:
-        ensure_model_supports_tools(
-            str(model), config.get("base_url") or config.get("openai_api_base")
-        )
-    except ModelCapabilityError as exc:
-        logger.error("Skipping MCP server '%s': %s", name, exc)
-        return None
-
-    subagent_name = str(agent_cfg.get("name") or "").strip()
-    subagent_description = str(agent_cfg.get("description") or "").strip()
-    instructions = str(agent_cfg.get("system_prompt") or "").strip()
-
-    if not subagent_name:
-        logger.error(
-            "Skipping MCP server '%s': missing required field agent.name", name
-        )
-        return None
-    if not subagent_description:
-        logger.error(
-            "Skipping MCP server '%s': missing required field agent.description", name
-        )
-        return None
-    if not instructions:
-        instructions = DEFAULT_AGENT_INSTRUCTIONS.format(name=name)
-
-    reasoning_effort = validate_reasoning_effort(
-        str(
-            agent_cfg.get("reasoning_effort")
-            or config.get("reasoning_effort")
-            or "medium"
-        )
-    )
-
-    raw_context_window = agent_cfg.get("context_window", config.get("context_window"))
-    context_window = (
-        int(str(raw_context_window)) if raw_context_window not in (None, "") else None
-    )
-
-    try:
-        # Use per-call MCP sessions instead of a shared long-lived session.
-        # This avoids task-affinity/cancellation issues (AnyIO cancel scope
-        # errors) when tools are invoked from nested subagents.
-        mcp_tools = await load_mcp_tools(
-            None,
-            connection=connection,  # type: ignore[arg-type]
-            server_name=name,
-            tool_name_prefix=False,
-        )
-        mcp_tools = _relax_const_fields(mcp_tools)
-    except Exception as exc:
-        logger.error("Failed to initialize MCP server '%s': %s", name, exc)
-        return None
-
-    logger.info("Initialized MCP server: %s", name)
-    warnings: list[str] = []
-    llm = create_ollama_chat_model(
-        model=str(model),
-        base_url=str(
-            config.get("base_url")
-            or config.get("openai_api_base")
-            or "http://localhost:11434"
-        ),
-        api_key=str(config.get("api_key") or config.get("openai_api_key") or "ollama"),
-        context_window=context_window,
-        reasoning_effort=reasoning_effort,
-        temperature=0,
-        warn_callback=warnings.append,
-    )
-    for warning in warnings:
-        logger.warning("MCP server '%s': %s", name, warning)
-    tool_names = ", ".join(t.name for t in mcp_tools)
-    enhanced_description = (
-        f"{subagent_description} (Tools: {tool_names})"
-        if mcp_tools
-        else subagent_description
-    )
-    return RunningMCPServer(
-        name=name,
-        subagent={
-            "name": subagent_name,
-            "description": enhanced_description,
-            "system_prompt": instructions,
-            "tools": mcp_tools,
-            "model": llm,
-        },
-        _closer=_noop_shutdown,
-    )
-
-
-async def initialize_mcp_servers(
-    config_path: Path | None = None,
-    *,
-    default_model: str | None = None,
-    base_url: str | None = None,
-    context_window: int | None = None,
-    reasoning_effort: ReasoningEffortValue = "medium",
-) -> list[RunningMCPServer]:
-    path = config_path or DEFAULT_MCP_CONFIG_PATH
-    if not path.exists():
-        return []
-    try:
-        data = json.loads(path.read_text(encoding="utf-8"))
-    except (json.JSONDecodeError, OSError) as exc:
-        logger.error("Failed to load MCP config %s: %s", path, exc)
-        return []
-
-    servers_cfg = data.get("mcpServers", {})
-    if not isinstance(servers_cfg, dict):
-        logger.warning("Invalid 'mcpServers' in config")
-        return []
-
-    servers: list[RunningMCPServer] = []
-    for name, cfg in servers_cfg.items():
-        if isinstance(cfg, dict):
-            if base_url and "base_url" not in cfg and "openai_api_base" not in cfg:
-                cfg = {**cfg, "base_url": base_url}
-            if context_window is not None and "context_window" not in cfg:
-                cfg = {**cfg, "context_window": context_window}
-            if reasoning_effort and "reasoning_effort" not in cfg:
-                cfg = {**cfg, "reasoning_effort": reasoning_effort}
-        if s := await _init_server(str(name), cfg, default_model):
-            servers.append(s)
-    return servers
-
-
-async def cleanup_mcp_servers(servers: list[RunningMCPServer]) -> None:
-    for server in servers:
-        await server.shutdown()
